Corrige rastreamento de arquivos build no adm_desktop

+++ adm_desktop/src/main/java/br/com/nearfood/repository/ProdutoRepository.java
@@ -26,7 +26,7 @@ public class ProdutoRepository {
     }
     
     public void atualizar (Produto produto){
-        delete(produto.getId());
+        delete((int) produto.getId());
         salvar(produto);
